+++ tests/Integration/Defer/DeferDBTest.php
@@ -13,15 +13,6 @@ use Tests\Utils\Models\User;
 
 class DeferDBTest extends DBTestCase
 {
-    use SetUpDefer;
-
-    protected function getEnvironmentSetUp($app)
-    {
-        parent::getEnvironmentSetUp($app);
-
-        $this->setUpDefer($app);
-    }
-
     protected function getPackageProviders($app)
     {
         return array_merge(
@@ -61,7 +52,7 @@ class DeferDBTest extends DBTestCase
             $queries++;
         });
 
-        $chunks = $this->getStreamedChunks(/** @lang GraphQL */ '
+        $chunks = $this->streamGraphQL(/** @lang GraphQL */ '
         {
             user {
                 email
@@ -117,7 +108,7 @@ class DeferDBTest extends DBTestCase
             $queries++;
         });
 
-        $chunks = $this->getStreamedChunks(/** @lang GraphQL */ '
+        $chunks = $this->streamGraphQL(/** @lang GraphQL */ '
         {
             user {
                 email
@@ -193,7 +184,7 @@ class DeferDBTest extends DBTestCase
             $queries++;
         });
 
-        $chunks = $this->getStreamedChunks(/** @lang GraphQL */ '
+        $chunks = $this->streamGraphQL(/** @lang GraphQL */ '
         {
             companies {
                 name